% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_iowa_data.R
\name{download_iowa_data}
\alias{download_iowa_data}
\title{Data set batch downloader}
\usage{
download_iowa_data(
  data_id,
  folder,
  data_name,
  total_of_rows = 10000,
  batch_size = 5000
)
}
\arguments{
\item{data_id}{A string. You get it from the data set page in. This can be found on the url of your data set of interest; for example the url for Iowa Fire Department Census data set is "https://data.iowa.gov/Emergency-Management/Iowa-Fire-Department-Census/hv43-6ksq/about_data" here the id is the value nex to the data set name on the right (i.e., "hv43-6ksq").}

\item{folder}{A string. The path to the folder you will save the data.}

\item{data_name}{A string. How your files will be called in your file system.}

\item{total_of_rows}{An integer. The size of the dataset, or the amount of rows you want to pull from the API.}

\item{batch_size}{An integer. The size of the batch (in rows) you want to pull, the default is 1M rows.}
}
\value{
Nothing. The function save the batches in .csv files in the folder you've chosen for.
}
\description{
\code{download_iowa_data()} downloads specified data from the Iowa Open Data portal (https://data.iowa.gov/).
}
\examples{
# Here we download the Iowa Fire Department Census data set.
# If you go to the dataset web page
# (https://data.iowa.gov/Emergency-Management/Iowa-Fire-Department-Census/hv43-6ksq/about_data),
# you will find that the total number of rows of this data set
# (to the date, 26-Jan-2024) is 738, so if we want the whole data
# set it is okay if we set `total_of_rows = 1000`, and `batch_size = 1000`.
download_iowa_data(data_id   = 'hv43-6ksq',
                   folder    = 'data',
                   data_name = 'fire_department_census',
                   total_of_rows = 1000,
                   batch_size    = 1000)
}
