% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_iowa_data.R
\name{read_iowa_data}
\alias{read_iowa_data}
\title{Read the data pulled from Iowa Open Data portal (https://data.iowa.gov/).}
\usage{
read_iowa_data(folder_path = "data", data_name)
}
\arguments{
\item{folder_path}{A string. The path to the folder you saved your data.}

\item{data_name}{A string. How your data set common files are called in your file system.}
}
\value{
A data frame. It reads into the memory the chunks of your data set downloaded in batches from the Iowa Open Data portal (https://data.iowa.gov/).
}
\description{
\code{read_iowa_data} reads the data pulled from the Iowa Open Data portal API with the function \code{download_iowa_data}
}
\examples{
# Read the fire_department_census data you downloaded with the download_iowa_data() function.
# The folder 'data' with the data is located in your R project working directory.
read_iowa_data(folder_path = 'data',
               data_name   = 'fire_department_census')
}
